# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing tasks and sandboxes.",
        "title": "Platform Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/schedules": {
            "get": {
                "description": "List all schedules for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "List schedules",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.scheduleResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new scheduled task.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "Create a schedule",
                "parameters": [
                    {
                        "description": "Create schedule request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.createScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_api.scheduleResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/schedules/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "Get a schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.scheduleResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "Update a schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.updateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.scheduleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "schedules"
                ],
                "summary": "Delete a schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/schedules/{id}/disable": {
            "post": {
                "tags": [
                    "schedules"
                ],
                "summary": "Disable a schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/schedules/{id}/enable": {
            "post": {
                "tags": [
                    "schedules"
                ],
                "summary": "Enable a schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/schedules/{id}/run": {
            "post": {
                "description": "Fires the schedule immediately and returns the task ID of the spawned run.",
                "tags": [
                    "schedules"
                ],
                "summary": "Trigger a manual run of a schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.scheduleRunResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/schedules/{id}/runs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "List past runs for a schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.runListItem"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.createScheduleRequest": {
            "type": "object",
            "required": [
                "cron_expr",
                "prompt"
            ],
            "properties": {
                "concurrency": {
                    "type": "integer"
                },
                "cron_expr": {
                    "type": "string"
                },
                "extra_env": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "git_url": {
                    "type": "string"
                },
                "prompt": {
                    "type": "string"
                },
                "run_at": {
                    "type": "string"
                },
                "timeout_secs": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.runListItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "error_msg": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.scheduleResponse": {
            "type": "object",
            "properties": {
                "concurrency": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "cron_expr": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "extra_env": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "git_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_run_at": {
                    "type": "string"
                },
                "next_run_at": {
                    "type": "string"
                },
                "prompt": {
                    "type": "string"
                },
                "run_at": {
                    "type": "string"
                },
                "timeout_secs": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "internal_api.scheduleRunResponse": {
            "type": "object",
            "properties": {
                "task_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.updateScheduleRequest": {
            "type": "object",
            "properties": {
                "concurrency": {
                    "type": "integer"
                },
                "cron_expr": {
                    "type": "string"
                },
                "extra_env": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "git_url": {
                    "type": "string"
                },
                "prompt": {
                    "type": "string"
                },
                "run_at": {
                    "type": "string"
                },
                "timeout_secs": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}